nessus: try to create a compressed zip file
Python's zipfile module creates uncompressed zip files by default, due
to a a bit legacy configuration setting, because Zlib support is optional.

Since it's not a problem in most modern Python builds, we try to create
the report zip file, with compression enabled, and if it fails, we roll
back to use the old no-compression setting.